requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
flask==2.2.2
flask-restx==1.3.2
requests==2.32.5
python-dotenv==1.2.1
werkzeug==2.2.2
